requirements.txt
flask==3.1.0
line-bot-sdk==3.5.1
anthropic==0.43.0
gspread==6.1.4
google-auth==2.37.0
gunicorn==23.0.0
